requirements.txt
fastapi==0.65.2
aiofiles==0.7.0
pyyaml
loguru
uvicorn